# app/api/movements/route.ts
import { NextRequest, NextResponse } from 'next/server'
import { getServerSession } from 'next-auth'
import { authOptions } from '@/foundation/lib/auth'
import { createMovementMaterial, getMovementMaterials } from '@/domains/movement-material/service'
import { createMovementMaterialSchema } from '@/domains/movement-material/validators'
import { z } from 'zod'

// GET /api/movements - List all movement materials
export async function GET() {
  try {
    const session = await getServerSession(authOptions)
    if (!session?.user) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
    }

    const movements = await getMovementMaterials()
    return NextResponse.json(movements)
  } catch (error) {
    console.error('List movements error:', error)
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 })
  }
}

// POST /api/movements - Create a new movement material
export async function POST(request: NextRequest) {
  try {
    const session = await getServerSession(authOptions)
    if (!session?.user) {
      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
    }

    const body = await request.json()
    const validated = createMovementMaterialSchema.parse(body)

    const movement = await createMovementMaterial(validated)
    return NextResponse.json(movement, { status: 201 })
  } catch (error) {
    if (error instanceof z.ZodError) {
      return NextResponse.json({ error: error.errors }, { status: 400 })
    }
    console.error('Create movement error:', error)
    return NextResponse.json({ error: 'Internal server error' }, { status: 500 })
  }
}